fixing type

+++ hello-spring-boot-rest/src/main/java/io/pivotal/hello/GreetingRepository.java
@@ -2,6 +2,6 @@ package io.pivotal.hello;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface GreetingRepository extends JpaRepository<Greeting, Long>{
+public interface GreetingRepository extends JpaRepository<Greeting, Integer>{
 	
 }
